requirements.txt
textual>=0.50.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
openai>=1.0.0
google-generativeai>=0.5.0
groq>=0.4.0
rich>=13.0.0
